requirements.txt
alembic==1.12.1
appnope==0.1.0
APScheduler==3.10.4
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.4
beautifulsoup4==4.12.3
blinker==1.4
cachelib==0.10.2
certifi==2024.6.2
cffi==1.15.1
charset-normalizer==3.4.0
click==8.1.7
decorator==4.3.0
exceptiongroup==1.2.2
flake8==5.0.4
Flask==2.2.5
Flask-Bcrypt==1.0.1
Flask-Cors==5.0.0
Flask-DebugToolbar==0.15.1
Flask-Migrate==4.0.7
Flask-Session==0.6.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.0.0
git-filter-repo==2.45.0
greenlet==3.0.3
idna==3.10
importlib-metadata==6.7.0
importlib-resources==5.12.0
iniconfig==2.0.0
ipython==7.34.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.4
Mako==1.2.4
MarkupSafe==2.1.5
matplotlib-inline==0.1.6
mccabe==0.7.0
neuralart==1.1.1
numpy==1.21.6
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.5.0
pluggy==1.2.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
Pygments==2.17.2
pytest==7.4.4
python-dotenv==0.21.1
pytz==2024.2
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.35
tomli==2.0.1
torch==1.13.1
traitlets==5.9.0
typing_extensions==4.7.1
tzlocal==5.1
urllib3==1.26.16
wcwidth==0.2.13
Werkzeug==2.2.3
WTForms==2.2.1
zipp==3.15.0
